+++ test/unit/parser.js
@@ -246,7 +246,7 @@
     assert.ok(fabric.parsePointsAttribute);
 
     var element = fabric.document.createElement('polygon');
-    element.setAttribute('points', '10,  12           20 ,22,  -0.52,0.001 2.3e2,2.3e-2, 10,-1     ');
+    element.setAttribute('points', '10,  12           20 ,22,  -0.52,0.001 2.3e2,2.3E-2, 10,-1     ');
 
     var actualPoints = fabric.parsePointsAttribute(element.getAttribute('points'));
 
@@ -304,9 +304,9 @@
     parsedValue = fabric.parseTransformAttribute(element.getAttribute('transform'));
     assert.deepEqual(parsedValue, [1,1.3820043381762832,0,1,0,0]);
 
-    element.setAttribute('transform', 'matrix(1,2,3,4,5,6)');
+    element.setAttribute('transform', 'matrix(1,2,3.3,-4,5E1,6e-1)');
     parsedValue = fabric.parseTransformAttribute(element.getAttribute('transform'));
-    assert.deepEqual(parsedValue, [1,2,3,4,5,6]);
+    assert.deepEqual(parsedValue, [1,2,3.3,-4,50,0.6]);
 
     element.setAttribute('transform', 'translate(21,31) translate(11,22)');
     parsedValue = fabric.parseTransformAttribute(element.getAttribute('transform'));
@@ -447,6 +447,28 @@
       assert.equal(objects[5].fillOpacity, 1,'first circle has fill-opacity 1');
       assert.equal(objects[6].opacity, 0.25, 'opacity is 0.25 for cls-3 * cls-4');
       assert.equal(objects[7].opacity, 0.5,'opacity is 0.5 from cls-3');
+      done();
+    });
+  });
+
+  QUnit.test('parseSVGFromString path fill-opacity with gradient', function(assert) {
+    var done = assert.async();
+    var string = '<?xml version="1.0" encoding="UTF-8"?>' +
+    '<svg version="1.2" baseProfile="tiny" xml:id="svg-root" width="300" height="400" ' +
+      'viewBox="0 0 300 400" xmlns="http://www.w3.org/2000/svg" ' +
+      'xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:xe="http://www.w3.org/2001/xml-events">' +
+        '<linearGradient id="red-to-red">' +
+          '<stop offset="0%" stop-color="#00ff00" stop-opacity="0.5"/>' +
+          '<stop offset="100%" stop-color="#ff0000"/>' +
+        '</linearGradient>' +
+        '<path d="M 0 0 l 100 0 l 0 100 l -100 0 z" fill="url(#red-to-red)" fill-opacity="0.5"/>' +
+    '</svg>';
+
+    fabric.loadSVGFromString(string, function(objects) {
+      assert.equal(objects[0].fill.colorStops[0].opacity, 0.5);
+      assert.equal(objects[0].fill.colorStops[0].color, 'rgb(255,0,0)');
+      assert.equal(objects[0].fill.colorStops[1].opacity, 0.25);
+      assert.equal(objects[0].fill.colorStops[1].color, 'rgb(0,255,0)');
       done();
     });
   });